2 этап(iter 65)

+++ main-service/src/main/java/ru/practicum/event/dto/EventShortDtoOut.java
@@ -19,6 +19,6 @@ public class EventShortDtoOut {
     @NotNull
     Boolean paid;
     String title;
-    Integer views;
+    Long views;
 
 }
